Fix issues with miusmatching Int and Long values in lookup maps

+++ src/main/kotlin/de/visualdigits/kaudiotagger/model/common/types/ReceivedAsTypes.kt
@@ -21,9 +21,9 @@ enum class ReceivedAsTypes(
         //The number of bytes used to hold the text encoding field size
         const val RECEIVED_AS_FIELD_SIZE: Int = 1
 
-        fun getValueToIdMap(): Map<String, Int> = entries.associate { e -> Pair(e.friendlyName, e.id) }
+        fun getValueToIdMap(): Map<String, Long> = entries.associate { e -> Pair(e.friendlyName, e.id.toLong()) }
 
-        fun getIdToValueMap(): Map<Int, String> = entries.associate { e -> Pair(e.id, e.friendlyName) }
+        fun getIdToValueMap(): Map<Long, String> = entries.associate { e -> Pair(e.id.toLong(), e.friendlyName) }
 
     }
 }
