+++ src/main/kotlin/de/visualdigits/kaudiotagger/model/common/types/SynchronisedLyricsContentType.kt
@@ -20,9 +20,8 @@ enum class SynchronisedLyricsContentType(
 
         const val CONTENT_KEY_FIELD_SIZE: Int = 1
 
-        fun getValueToIdMap(): Map<String, Int> = entries.associate { e -> Pair(e.friendlyName, e.id) }
+        fun getValueToIdMap(): Map<String, Long> = entries.associate { e -> Pair(e.friendlyName, e.id.toLong()) }
 
-        fun getIdToValueMap(): Map<Int, String> = entries.associate { e -> Pair(e.id, e.friendlyName) }
-
+        fun getIdToValueMap(): Map<Long, String> = entries.associate { e -> Pair(e.id.toLong(), e.friendlyName) }
     }
 }
